Update index.html
Changed the logic of B2C and B2B.
Previously, I thought if a seller sells a used car under margin scheme, the purchaser will lose the option to sell it further in the margin scheme. But it does not matter whether or not the purchaser gets the car under margin scheme, the purchaser has to decide to avail the ITC in such purchase or not. If the purchaser can be able to sell the used car if he does not avail ITC on the purchased used car.